Update buildscript to support x64 arch We use std::filesystem in this place
