+++ extension/filesystemwatcher.cpp
@@ -8,7 +8,7 @@
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License, version 3.0, as published by the
  * Free Software Foundation.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
@@ -28,26 +28,21 @@
  */
 
 #include "filesystemwatcher.h"
-#include "smsdk_ext.h"
+#include <cstring>
+
+#ifdef __linux__
+#else
 #include <locale>
 #include <codecvt>
-#include <am-string.h>
-#include <os/am-fsutil.h>
-#include <os/am-path.h>
-
-#ifdef KE_LINUX
-#include <sys/eventfd.h>
-#include <poll.h>
 #endif
 
+#include "smsdk_ext.h"
 #include "helpers.h"
 
-FileSystemWatcher::FileSystemWatcher(const std::string &relPath) :
-	m_watching(false),
-	m_includeSubdirectories(false),
-	m_notifyFilter(FSW_NOTIFY_NONE),
-	m_bufferSize(8192),
-	m_retryInterval(1000),
+namespace fs = std::filesystem;
+
+SMDirectoryWatcher::SMDirectoryWatcher(const std::string& path) :
+	DirectoryWatcher((fs::path(g_pSM->GetGamePath()) / path).lexically_normal()),
 	m_Handle(0),
 	m_owningContext(nullptr),
 	m_onStarted(nullptr),
@@ -55,122 +50,19 @@ FileSystemWatcher::FileSystemWatcher(const std::string &relPath) :
 	m_onCreated(nullptr),
 	m_onDeleted(nullptr),
 	m_onModified(nullptr),
-	m_onRenamed(nullptr),
-#ifdef KE_WINDOWS
-	m_threadCancelEventHandle(nullptr),
-#elif defined KE_LINUX
-	m_threadCancelEventHandle(-1),
-#endif
-	m_thread(),
-	m_threadRunning(false),
-	m_processingEvents(false)
+	m_onRenamed(nullptr)
 {
-	char buffer[PLATFORM_MAX_PATH];
-	ke::path::Format(buffer, sizeof(buffer), relPath.c_str());
-	m_relPath = buffer;
-
-	if (!m_relPath.empty() && (m_relPath.back() != '/' && m_relPath.back() != '\\'))
-	{
-#ifdef KE_WINDOWS
-		m_relPath.push_back('\\');
-#else
-		m_relPath.push_back('/');
-#endif
-	}
-
-	g_pSM->BuildPath(Path_Game, buffer, sizeof(buffer), "%s", m_relPath.c_str());
-
-	m_path = buffer;
+	m_relPath = fs::path(path).lexically_normal();
 }
 
-FileSystemWatcher::~FileSystemWatcher()
+size_t SMDirectoryWatcher::GetPath(char* buffer, size_t bufferSize)
 {
-	Stop();
-
-#ifdef KE_WINDOWS
-	if (m_threadCancelEventHandle)
-	{
-		CloseHandle(m_threadCancelEventHandle);
-	}
-#elif defined KE_LINUX
-	if (fcntl(m_threadCancelEventHandle, F_GETFD) != -1)
-	{
-		close(m_threadCancelEventHandle);	
-	}
-#endif
+	std::strncpy(buffer, m_relPath.string().c_str(), bufferSize - 1);
+	buffer[bufferSize - 1] = '\0';
+	return strlen(buffer) + 1;
 }
 
-bool FileSystemWatcher::Start()
-{
-	if (IsWatching())
-	{
-		return true;
-	}
-
-	std::unique_ptr<ThreadConfig> config = std::make_unique<ThreadConfig>();
-	config->root_path = m_path;
-	config->includeSubdirectories = m_includeSubdirectories;
-	config->notifyFilters = m_notifyFilter;
-	config->bufferSize = m_bufferSize;
-	config->retryInterval = m_retryInterval;
-
-#ifdef KE_WINDOWS
-	if (!m_threadCancelEventHandle)
-	{
-		m_threadCancelEventHandle = CreateEvent(nullptr, TRUE, FALSE, nullptr);
-	}
-	else
-	{
-		ResetEvent(m_threadCancelEventHandle);
-	}
-#elif defined KE_LINUX
-	if (fcntl(m_threadCancelEventHandle, F_GETFD) == -1)
-	{
-		m_threadCancelEventHandle = eventfd(0, 0);
-	}
-	else
-	{
-		uint64_t u;
-		read(m_threadCancelEventHandle, &u, sizeof(u));
-	}
-#endif
-
-	m_watching = true;
-	m_thread = std::thread(&FileSystemWatcher::ThreadProc, this, std::move(config));
-
-	return true;
-}
-
-size_t FileSystemWatcher::GetAbsolutePath(char* buffer, size_t bufferSize)
-{
-	return ke::SafeStrcpy(buffer, bufferSize, m_path.c_str());
-}
-
-size_t FileSystemWatcher::GetRelativePath(char* buffer, size_t bufferSize)
-{
-	return ke::SafeStrcpy(buffer, bufferSize, m_relPath.c_str());
-}
-
-void FileSystemWatcher::Stop()
-{
-	if (!IsWatching())
-	{
-		return;
-	}
-
-	m_watching = false;
-
-	RequestCancelThread();
-
-	if (m_thread.joinable())
-	{
-		m_thread.join();
-	}
-
-	ProcessEvents();
-}
-
-void FileSystemWatcher::OnGameFrame(bool simulating)
+void SMDirectoryWatcher::OnGameFrame(bool simulating)
 {
 	if (IsWatching())
 	{
@@ -178,7 +70,7 @@ void FileSystemWatcher::OnGameFrame(bool simulating)
 	}
 }
 
-void FileSystemWatcher::OnPluginUnloaded(SourceMod::IPlugin* plugin)
+void SMDirectoryWatcher::OnPluginUnloaded(SourceMod::IPlugin* plugin)
 {
 	IPluginContext* context = plugin->GetBaseContext();
 
@@ -219,758 +111,82 @@ void FileSystemWatcher::OnPluginUnloaded(SourceMod::IPlugin* plugin)
 	}
 }
 
-bool FileSystemWatcher::IsThreadRunning()
+void SMDirectoryWatcher::OnProcessEvent(const NotifyEvent& event)
 {
-	std::lock_guard<std::mutex> lock(m_threadRunningMutex);
-	return m_threadRunning;
-}
-
-void FileSystemWatcher::SetThreadRunning(bool state)
-{
-	std::lock_guard<std::mutex> lock(m_threadRunningMutex);
-	m_threadRunning = state;
-}
-
-void FileSystemWatcher::RequestCancelThread()
-{
-#ifdef KE_WINDOWS
-	SetEvent(m_threadCancelEventHandle);
-#elif defined KE_LINUX
-	uint64_t u = 1;
-	write(m_threadCancelEventHandle, &u, sizeof(u));
-#endif
-}
-
-void FileSystemWatcher::ThreadProc(std::unique_ptr<ThreadConfig> config)
-{
-	class ThreadData
+	switch (event.type)
 	{
-	public:
-#ifdef KE_WINDOWS
-		HANDLE directory;
-		HANDLE changeEvent;
-#elif defined KE_LINUX
-		int fd;
-		int root_wd;
-		std::map<int, std::string> wd;
-		uint32_t mask;
-#endif
-
-		ThreadData()
+		case NotifyEvent::NotifyEventType::FILESYSTEM:
 		{
-#ifdef KE_WINDOWS
-			directory = INVALID_HANDLE_VALUE;
-			changeEvent = nullptr;
-#elif defined KE_LINUX
-			fd = -1;
-			root_wd = -1;
-			mask = 0;
-#endif
-		}
-
-		~ThreadData()
-		{
-#ifdef KE_WINDOWS
-		if (changeEvent && changeEvent != INVALID_HANDLE_VALUE)
-		{
-			CloseHandle(changeEvent);
-		}
-
-		if (directory && directory != INVALID_HANDLE_VALUE)
-		{
-			CloseHandle(directory);
-		}
-#elif defined KE_LINUX
-		if (fcntl(fd, F_GETFD) != -1)
-		{
-			for (auto it = wd.begin(); it != wd.end(); it++)
+			if (event.flags & FSW_NOTIFY_CREATED)
 			{
-				inotify_rm_watch(fd, it->first);
-			}
-
-			close(fd);
-		}
-#endif
-		}
-
-#if defined KE_LINUX
-		int AddWatch(const std::unique_ptr<ThreadConfig> &config, const std::string &path, uint32_t _mask = 0)
-		{
-			int _wd = -1;
-			_mask |= mask;
-			
-			std::string absPath(config->root_path);
-			absPath.append(path);
-
-			_wd = inotify_add_watch(fd, absPath.c_str(), _mask);
-			if (_wd != -1)
-			{
-				auto it = wd.find(_wd);
-				if (it != wd.end())
+				if (m_onCreated && m_onCreated->IsRunnable())
 				{
-					wd.erase(it);
-				}
-				
-				wd.emplace(_wd, path);
-			}
-
-			if (config->includeSubdirectories)
-			{
-				DIR *dir;
-				dirent *ent;
-				if ((dir = opendir(absPath.c_str())) != nullptr) 
-				{
-					while ((ent = readdir(dir)) != nullptr)
-					{
-						if ((ent->d_type == DT_LNK || ent->d_type == DT_DIR) && ent->d_name[0] != '.')
-						{
-							if (ent->d_type == DT_LNK)
-							{
-								std::string linkPath(absPath);
-								linkPath.append(ent->d_name);
-
-								struct stat targetInfo;
-								if (stat(linkPath.c_str(), &targetInfo) == -1)
-								{
-									continue;
-								}
-
-								if (!S_ISDIR(targetInfo.st_mode))
-								{
-									continue;
-								}
-							}
-
-							std::string dirPath(path);
-							dirPath.append(ent->d_name);
-							dirPath.push_back('/');
-
-							AddWatch(config, dirPath, mask);
-						}
-					}
-
-					closedir(dir);
+					m_onCreated->PushCell(m_Handle);
+					m_onCreated->PushString(event.path.c_str());
+					m_onCreated->Execute(nullptr);
 				}
 			}
 
-			return _wd;
-		}
-#endif
-	};
-
-	bool isRunning = true;
-
-	while (isRunning)
-	{
-		SetThreadRunning(true);
-
-		bool didStart = false;
-
-		std::unique_ptr<ThreadData> data = std::make_unique<ThreadData>();
-		std::unique_ptr<char[]> buffer(new char[config->bufferSize]);
-
-		if (!ke::file::IsDirectory(config->root_path.c_str()))
-		{
-			goto terminate;
-		}
-
-#ifdef KE_WINDOWS
-		DWORD dwNotifyFilter = 0;
-
-		if (config->notifyFilters & (FSW_NOTIFY_RENAMED | FSW_NOTIFY_CREATED | FSW_NOTIFY_DELETED))
-		{
-			dwNotifyFilter |= FILE_NOTIFY_CHANGE_FILE_NAME;
-			dwNotifyFilter |= FILE_NOTIFY_CHANGE_DIR_NAME;
-		}
-		
-		if (config->notifyFilters & FSW_NOTIFY_MODIFIED)
-		{
-			dwNotifyFilter |= FILE_NOTIFY_CHANGE_LAST_WRITE;
-		}
-		
-		data->directory = CreateFileA(
-			config->root_path.c_str(),
-			FILE_LIST_DIRECTORY | GENERIC_READ,
-			FILE_SHARE_READ | FILE_SHARE_WRITE,
-			nullptr,
-			OPEN_EXISTING,
-			FILE_FLAG_BACKUP_SEMANTICS | FILE_FLAG_OVERLAPPED,
-			nullptr);
-
-		if (data->directory == INVALID_HANDLE_VALUE)
-		{
-			goto terminate;
-		}
-#elif defined KE_LINUX
-		data->fd = inotify_init1(IN_NONBLOCK);
-		if (data->fd == -1)
-		{
-			goto terminate;
-		}
-
-		data->mask = 0;
-
-		if (config->includeSubdirectories)
-		{
-			data->mask |= (IN_CREATE | IN_MOVE);
-		}
-		else
-		{
-			if (config->notifyFilters & FSW_NOTIFY_CREATED)
+			if (event.flags & FSW_NOTIFY_DELETED)
 			{
-				data->mask |= IN_CREATE;
-			}
-
-			if (config->notifyFilters & FSW_NOTIFY_RENAMED)
-			{
-				data->mask |= IN_MOVE;
-			}
-		}
-
-		if (config->notifyFilters & FSW_NOTIFY_DELETED)
-		{
-			data->mask |= IN_DELETE;
-		}
-
-		if (config->notifyFilters & FSW_NOTIFY_MODIFIED)
-		{
-			data->mask |= IN_CLOSE_WRITE;
-		}
-
-		data->root_wd = data->AddWatch(config, "", IN_DELETE_SELF | IN_MOVE_SELF);
-		if (data->root_wd == -1)
-		{
-			goto terminate;
-		}
-#endif
-
-		didStart = true;
-
-		{
-			std::lock_guard<std::mutex> lock(m_changeEventsMutex);
-
-			auto ev = std::make_unique<NotifyEvent>();
-			ev->type = NotifyEvent::NotifyEventType::START;
-			m_changeEvents.push(std::move(ev));
-		}
-
-#ifdef KE_WINDOWS
-		data->changeEvent = CreateEvent(nullptr, TRUE, FALSE, nullptr);
-
-		HANDLE waitHandles[2];
-		waitHandles[0] = data->changeEvent;
-		waitHandles[1] = m_threadCancelEventHandle;
-
-		OVERLAPPED overlapped;
-		ZeroMemory(&overlapped, sizeof overlapped);
-		overlapped.hEvent = data->changeEvent;
-
-		while (true)
-		{
-			if (!ReadDirectoryChangesExW(data->directory,
-				buffer.get(),
-				config->bufferSize,
-				config->includeSubdirectories,
-				dwNotifyFilter,
-				nullptr,
-				&overlapped,
-				nullptr,
-				ReadDirectoryNotifyExtendedInformation))
-			{
-				goto terminate;
-			}
-
-			switch (WaitForMultipleObjects(2, waitHandles, FALSE, INFINITE))
-			{
-				case WAIT_OBJECT_0:
+				if (m_onDeleted && m_onDeleted->IsRunnable())
 				{
-					DWORD dwBytes = 0;
-					if (!GetOverlappedResult(data->directory, &overlapped, &dwBytes, TRUE))
-					{
-						goto terminate;
-					}
-
-					ResetEvent(data->changeEvent);
-
-					if (dwBytes == 0)
-					{
-						break;
-					}
-
-					std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> converter;
-					std::vector<std::unique_ptr<NotifyEvent>> queuedEvents;
-
-					char *p = buffer.get();
-					for (;;)
-					{
-						FILE_NOTIFY_EXTENDED_INFORMATION* info = reinterpret_cast<FILE_NOTIFY_EXTENDED_INFORMATION*>(p);
-
-						std::wstring fileName(info->FileName, info->FileNameLength / sizeof(wchar_t));
-
-						switch (info->Action)
-						{
-							case FILE_ACTION_ADDED:
-							{
-								if (config->notifyFilters & (FSW_NOTIFY_CREATED | FSW_NOTIFY_RENAMED))
-								{
-									std::string path = converter.to_bytes(fileName);
-
-									if (!queuedEvents.empty())
-									{
-										bool isRenameEvent = false;
-
-										for (auto it = queuedEvents.rbegin(); it != queuedEvents.rend(); it++)
-										{
-											auto &change = *it;
-
-											if (change->flags == FSW_NOTIFY_DELETED)
-											{
-												std::string me = GetFileNameFromPath(path);
-												std::string them = GetFileNameFromPath(change->path);
-
-												if (me == them)
-												{
-													change->flags = FSW_NOTIFY_RENAMED;
-													change->lastPath = change->path;
-													change->path = path;
-
-													isRenameEvent = true;
-													break;
-												}
-											}
-										}
-
-										if (isRenameEvent)
-										{
-											break;
-										}
-									}
-
-									auto change = std::make_unique<NotifyEvent>();
-									change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-									change->flags = FSW_NOTIFY_CREATED;
-									change->path = path;
-
-									queuedEvents.push_back(std::move(change));
-								}
-
-								break;
-							}
-							case FILE_ACTION_REMOVED:
-							{
-								if (config->notifyFilters & (FSW_NOTIFY_DELETED | FSW_NOTIFY_RENAMED))
-								{
-									auto change = std::make_unique<NotifyEvent>();
-									change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-									change->flags = FSW_NOTIFY_DELETED;
-									change->path = converter.to_bytes(fileName);
-
-									queuedEvents.push_back(std::move(change));
-								}
-								
-								break;
-							}
-							case FILE_ACTION_MODIFIED:
-							{
-								if (config->notifyFilters & FSW_NOTIFY_MODIFIED)
-								{
-									if (info->FileAttributes & FILE_ATTRIBUTE_DIRECTORY)
-									{
-										break;
-									}
-
-									std::string path = converter.to_bytes(fileName);
-
-									std::string absPath(config->root_path);
-									absPath.append(path);
-
-									auto change = std::make_unique<NotifyEvent>();
-									change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-									change->flags = FSW_NOTIFY_MODIFIED;
-									change->path = path;
-
-									queuedEvents.push_back(std::move(change));
-								}
-
-								break;
-							}
-							case FILE_ACTION_RENAMED_OLD_NAME:
-							{
-								if (config->notifyFilters & FSW_NOTIFY_RENAMED)
-								{
-									auto change = std::make_unique<NotifyEvent>();
-									change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-									change->flags = FSW_NOTIFY_RENAMED;
-									change->lastPath = converter.to_bytes(fileName);
-
-									queuedEvents.push_back(std::move(change));
-								}
-								
-								break;
-							}
-							case FILE_ACTION_RENAMED_NEW_NAME:
-							{
-								if (config->notifyFilters & FSW_NOTIFY_RENAMED)
-								{
-									auto &change = queuedEvents.back();
-									change->path = converter.to_bytes(fileName);
-								}
-								
-								break;
-							}
-						}
-
-						if (!info->NextEntryOffset)
-						{
-							break;
-						}
-
-						p += info->NextEntryOffset;
-					}
-
-					std::lock_guard<std::mutex> lock(m_changeEventsMutex);
-
-					for (auto it = queuedEvents.begin(); it != queuedEvents.end(); it++)
-					{
-						auto &change = *it;
-						if (change->flags & config->notifyFilters)
-						{
-							m_changeEvents.push(std::move(*it));
-						}
-					}
-
-					break;
-				}
-				case WAIT_OBJECT_0 + 1:
-				{
-					isRunning = false;
-					goto terminate;
-				}
-				default:
-				{
-					goto terminate;
+					m_onDeleted->PushCell(m_Handle);
+					m_onDeleted->PushString(event.path.c_str());
+					m_onDeleted->Execute(nullptr);
 				}
 			}
-		}
-#elif defined KE_LINUX
-		pollfd fds[2];
 
-		fds[0].fd = data->fd;
-		fds[0].events = POLLIN;
-
-		fds[1].fd = m_threadCancelEventHandle;
-		fds[1].events = POLLIN;
-
-		while (true)
-		{
-			if (poll(fds, 2, -1) > 0)
+			if (event.flags & FSW_NOTIFY_MODIFIED)
 			{
-				if (fds[1].revents & POLLIN)
+				if (m_onModified && m_onModified->IsRunnable())
 				{
-					isRunning = false;
-					goto terminate;
-				}
-
-				if (fds[0].revents & POLLIN)
-				{
-					std::vector<std::unique_ptr<NotifyEvent>> queuedEvents;
-
-					for (;;)
-					{
-						ssize_t len = read(data->fd, buffer.get(), config->bufferSize);
-						
-						if (len == -1 && errno != EAGAIN) 
-						{
-							goto terminate;
-						}
-
-						if (len <= 0)
-						{
-							break;
-						}
-
-						const inotify_event* event;
-						for (char* p = buffer.get(); p < buffer.get() + len; p += sizeof(inotify_event) + event->len)
-						{
-							event = (const inotify_event*)p;
-
-							if (event->mask & (IN_MOVE_SELF | IN_DELETE_SELF | IN_IGNORED))
-							{
-								auto it = data->wd.find(event->wd);
-								if (it != data->wd.end())
-								{
-									if (event->mask & IN_MOVE_SELF)
-									{
-										inotify_rm_watch(data->fd, event->wd);
-									}
-
-									data->wd.erase(it);
-								}
-
-								if (event->wd == data->root_wd)
-								{
-									goto terminate;
-								}
-								else
-								{
-									continue;
-								}
-							}
-
-							std::string &baseRelPath = data->wd.at(event->wd);
-
-							if (event->mask & (IN_CREATE | IN_MOVED_TO))
-							{
-								if (config->includeSubdirectories && (event->mask & IN_ISDIR))
-								{
-									std::string relPath(baseRelPath);
-									relPath.append(event->name);
-									relPath.push_back('/');
-
-									data->AddWatch(config, relPath);
-								}
-
-								if (event->mask & IN_MOVED_TO)
-								{
-									bool foundRenamedEvent = false;
-
-									for (auto it = queuedEvents.rbegin(); it != queuedEvents.rend(); it++)
-									{
-										auto &change = *it;
-										if (change->cookie == event->cookie)
-										{
-											change->flags = FSW_NOTIFY_RENAMED;
-											change->cookie = 0;
-											change->lastPath = change->path;
-											change->path = baseRelPath;
-											change->path.append(event->name);
-
-											foundRenamedEvent = true;
-											break;
-										}
-									}
-
-									if (foundRenamedEvent)
-									{
-										continue;
-									}
-								}
-
-								std::unique_ptr<NotifyEvent> change = std::make_unique<NotifyEvent>();
-								change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-								change->flags = FSW_NOTIFY_CREATED;
-								change->cookie = event->cookie;
-								change->path = baseRelPath;
-								change->path.append(event->name);
-
-								queuedEvents.push_back(std::move(change));
-							}
-
-							if (event->mask & (IN_DELETE | IN_MOVED_FROM))
-							{
-								if (event->mask & IN_MOVED_FROM)
-								{
-									bool foundRenamedEvent = false;
-
-									for (auto it = queuedEvents.rbegin(); it != queuedEvents.rend(); it++)
-									{
-										auto &change = *it;
-										if (change->cookie == event->cookie)
-										{
-											change->flags = FSW_NOTIFY_RENAMED;
-											change->cookie = 0;
-											change->lastPath = change->path;
-											change->path = baseRelPath;
-											change->path.append(event->name);
-
-											foundRenamedEvent = true;
-											break;
-										}
-									}
-
-									if (foundRenamedEvent)
-									{
-										continue;
-									}
-								}
-
-								std::unique_ptr<NotifyEvent> change = std::make_unique<NotifyEvent>();
-								change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-								change->flags = FSW_NOTIFY_DELETED;
-								change->cookie = event->cookie;
-								change->path = baseRelPath;
-								change->path.append(event->name);
-
-								queuedEvents.push_back(std::move(change));
-							}
-
-							if (event->mask & IN_CLOSE_WRITE)
-							{
-								if (config->notifyFilters & FSW_NOTIFY_MODIFIED)
-								{
-									std::unique_ptr<NotifyEvent> change = std::make_unique<NotifyEvent>();
-									change->type = NotifyEvent::NotifyEventType::FILESYSTEM;
-									change->flags = FSW_NOTIFY_MODIFIED;
-									change->cookie = event->cookie;
-									change->path = baseRelPath;
-									change->path.append(event->name);
-
-									queuedEvents.push_back(std::move(change));
-								}
-							}
-						}
-					}
-
-					std::lock_guard<std::mutex> lock(m_changeEventsMutex);
-
-					for (auto it = queuedEvents.begin(); it != queuedEvents.end(); it++)
-					{
-						auto &change = *it;
-						if (change->flags & config->notifyFilters)
-						{
-							m_changeEvents.push(std::move(*it));
-						}
-					}
+					m_onModified->PushCell(m_Handle);
+					m_onModified->PushString(event.path.c_str());
+					m_onModified->Execute(nullptr);
 				}
 			}
-		}
-#endif
 
-terminate:
-		SetThreadRunning(false);
-
-		if (didStart)
-		{
-			std::lock_guard<std::mutex> lock(m_changeEventsMutex);
-
-			auto ev = std::make_unique<NotifyEvent>();
-			ev->type = NotifyEvent::NotifyEventType::EXIT;
-			m_changeEvents.push(std::move(ev));
-		}
-
-		if (isRunning)
-		{
-#ifdef KE_WINDOWS
-			switch (WaitForSingleObject(m_threadCancelEventHandle, config->retryInterval))
+			if (event.flags & FSW_NOTIFY_RENAMED)
 			{
-				case WAIT_OBJECT_0:
+				if (m_onRenamed && m_onRenamed->IsRunnable())
 				{
-					isRunning = false;
-					break;
+					m_onRenamed->PushCell(m_Handle);
+					m_onRenamed->PushString(event.lastPath.c_str());
+					m_onRenamed->PushString(event.path.c_str());
+					m_onRenamed->Execute(nullptr);
 				}
 			}
-#elif defined KE_LINUX
-			pollfd retry_fds[1];
 
-			retry_fds[0].fd = m_threadCancelEventHandle;
-			retry_fds[0].events = POLLIN;
-
-			if (poll(retry_fds, 1, config->retryInterval) > 0)
+			break;
+		}
+		case NotifyEvent::NotifyEventType::START:
+		{
+			if (m_onStarted && m_onStarted->IsRunnable())
 			{
-				if (retry_fds[0].revents & POLLIN)
-				{
-					isRunning = false;
-				}
+				m_onStarted->PushCell(m_Handle);
+				m_onStarted->Execute(nullptr);
 			}
-#endif
+
+			break;
+		}
+		case NotifyEvent::NotifyEventType::EXIT:
+		{
+			if (m_onStopped && m_onStopped->IsRunnable())
+			{
+				m_onStopped->PushCell(m_Handle);
+				m_onStopped->Execute(nullptr);
+			}
+
+			break;
 		}
 	}
 }
 
-void FileSystemWatcher::ProcessEvents()
-{
-	if (m_processingEvents)
-	{
-		return;
-	}
+SMDirectoryWatcherManager g_FileSystemWatchers;
+SourceMod::HandleType_t SMDirectoryWatcherManager::m_HandleType(0);
 
-	std::lock_guard<std::mutex> lock(m_changeEventsMutex);
-
-	m_processingEvents = true;
-
-	while (!m_changeEvents.empty())
-	{
-		std::unique_ptr<NotifyEvent> &front = m_changeEvents.front();
-
-		switch (front->type)
-		{
-			case NotifyEvent::NotifyEventType::FILESYSTEM:
-			{
-				if (front->flags & FSW_NOTIFY_CREATED)
-				{
-					if (m_onCreated && m_onCreated->IsRunnable())
-					{
-						m_onCreated->PushCell(m_Handle);
-						m_onCreated->PushString(front->path.c_str());
-						m_onCreated->Execute(nullptr);
-					}
-				}
-
-				if (front->flags & FSW_NOTIFY_DELETED)
-				{
-					if (m_onDeleted && m_onDeleted->IsRunnable())
-					{
-						m_onDeleted->PushCell(m_Handle);
-						m_onDeleted->PushString(front->path.c_str());
-						m_onDeleted->Execute(nullptr);
-					}
-				}
-
-				if (front->flags & FSW_NOTIFY_MODIFIED)
-				{
-					if (m_onModified && m_onModified->IsRunnable())
-					{
-						m_onModified->PushCell(m_Handle);
-						m_onModified->PushString(front->path.c_str());
-						m_onModified->Execute(nullptr);
-					}
-				}
-
-				if (front->flags & FSW_NOTIFY_RENAMED)
-				{
-					if (m_onRenamed && m_onRenamed->IsRunnable())
-					{
-						m_onRenamed->PushCell(m_Handle);
-						m_onRenamed->PushString(front->lastPath.c_str());
-						m_onRenamed->PushString(front->path.c_str());
-						m_onRenamed->Execute(nullptr);
-					}
-				}
-
-				break;
-			}
-			case NotifyEvent::NotifyEventType::START:
-			{
-				if (m_onStarted && m_onStarted->IsRunnable())
-				{
-					m_onStarted->PushCell(m_Handle);
-					m_onStarted->Execute(nullptr);
-				}
-
-				break;
-			}
-			case NotifyEvent::NotifyEventType::EXIT:
-			{
-				if (m_onStopped && m_onStopped->IsRunnable())
-				{
-					m_onStopped->PushCell(m_Handle);
-					m_onStopped->Execute(nullptr);
-				}
-
-				break;
-			}
-		}
-
-		m_changeEvents.pop();
-	}
-
-	m_processingEvents = false;
-}
-
-FileSystemWatcherManager g_FileSystemWatchers;
-SourceMod::HandleType_t FileSystemWatcherManager::m_HandleType(0);
-
-FileSystemWatcherManager::FileSystemWatcherManager()
+SMDirectoryWatcherManager::SMDirectoryWatcherManager()
 {
 }
 
@@ -979,12 +195,12 @@ static void GameFrameHook(bool simulating)
 	g_FileSystemWatchers.OnGameFrame(simulating);
 }
 
-bool FileSystemWatcherManager::SDK_OnLoad(char* error, int errorSize)
+bool SMDirectoryWatcherManager::SDK_OnLoad(char* error, int errorSize)
 {
 	m_HandleType = g_pHandleSys->CreateType("FileSystemWatcher", this, 0, nullptr, nullptr, myself->GetIdentity(), nullptr);
 	if (!m_HandleType)
 	{
-		ke::SafeStrcpy(error, errorSize, "Failed to create FileSystemWatcher handle type.");
+		std::snprintf(error, errorSize, "Failed to create FileSystemWatcher handle type.");
 		return false;
 	}
 
@@ -997,7 +213,7 @@ bool FileSystemWatcherManager::SDK_OnLoad(char* error, int errorSize)
 	return true;
 }
 
-void FileSystemWatcherManager::SDK_OnUnload()
+void SMDirectoryWatcherManager::SDK_OnUnload()
 {
 	smutils->RemoveGameFrameHook(&GameFrameHook);
 
@@ -1009,7 +225,7 @@ void FileSystemWatcherManager::SDK_OnUnload()
 	}
 }
 
-void FileSystemWatcherManager::OnGameFrame(bool simulating)
+void SMDirectoryWatcherManager::OnGameFrame(bool simulating)
 {
 	for (auto it = m_watchers.begin(); it != m_watchers.end(); it++)
 	{
@@ -1017,29 +233,29 @@ void FileSystemWatcherManager::OnGameFrame(bool simulating)
 	}
 }
 
-SourceMod::Handle_t FileSystemWatcherManager::CreateWatcher(SourcePawn::IPluginContext* context, const std::string &path)
+SourceMod::Handle_t SMDirectoryWatcherManager::CreateWatcher(SourcePawn::IPluginContext* context, const std::string &path)
 {
-	FileSystemWatcher* watcher = new FileSystemWatcher(path);
+	SMDirectoryWatcher* watcher = new SMDirectoryWatcher(path);
 	watcher->m_owningContext = context;
 	watcher->m_Handle = handlesys->CreateHandle(m_HandleType, watcher, context->GetIdentity(), myself->GetIdentity(), nullptr);
 	m_watchers.push_back(watcher);
 	return watcher->m_Handle;
 }
 
-FileSystemWatcher* FileSystemWatcherManager::GetWatcher(SourceMod::Handle_t handle)
+SMDirectoryWatcher* SMDirectoryWatcherManager::GetWatcher(SourceMod::Handle_t handle)
 {
-	FileSystemWatcher* watcher = nullptr;
+	SMDirectoryWatcher* watcher = nullptr;
 	HandleSecurity sec(nullptr, myself->GetIdentity());
 
 	SourceMod::HandleError err = g_pHandleSys->ReadHandle(handle, m_HandleType, &sec, (void**)(&watcher));
 	return (err == HandleError_None) ? watcher : nullptr;
 }
 
-void FileSystemWatcherManager::OnHandleDestroy(SourceMod::HandleType_t type, void *object)
+void SMDirectoryWatcherManager::OnHandleDestroy(SourceMod::HandleType_t type, void *object)
 {
 	if (type == m_HandleType)
 	{
-		FileSystemWatcher* watcher = (FileSystemWatcher*)object;
+		SMDirectoryWatcher* watcher = (SMDirectoryWatcher*)object;
 
 		for (auto it = m_watchers.begin(); it != m_watchers.end();)
 		{
@@ -1057,7 +273,7 @@ void FileSystemWatcherManager::OnHandleDestroy(SourceMod::HandleType_t type, voi
 	}
 }
 
-void FileSystemWatcherManager::OnPluginUnloaded(SourceMod::IPlugin* plugin)
+void SMDirectoryWatcherManager::OnPluginUnloaded(SourceMod::IPlugin* plugin)
 {
 	for (auto it = m_watchers.begin(); it != m_watchers.end(); it++)
 	{
@@ -1077,7 +293,7 @@ cell_t Native_FileSystemWatcher(SourcePawn::IPluginContext *context, const cell_
 
 cell_t Native_FileSystemWatcher_IncludeSubdirectoriesGet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1089,7 +305,7 @@ cell_t Native_FileSystemWatcher_IncludeSubdirectoriesGet(SourcePawn::IPluginCont
 
 cell_t Native_FileSystemWatcher_IncludeSubdirectoriesSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1102,7 +318,7 @@ cell_t Native_FileSystemWatcher_IncludeSubdirectoriesSet(SourcePawn::IPluginCont
 
 cell_t Native_FileSystemWatcher_NotifyFilterGet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1114,20 +330,20 @@ cell_t Native_FileSystemWatcher_NotifyFilterGet(SourcePawn::IPluginContext *cont
 
 cell_t Native_FileSystemWatcher_NotifyFilterSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
 		return 0;
 	}
 
-	watcher->m_notifyFilter = (FileSystemWatcher::NotifyFilters)params[2];
+	watcher->m_notifyFilter = (DirectoryWatcher::NotifyFilters)params[2];
 	return 0;
 }
 
 cell_t Native_FileSystemWatcher_RetryIntervalGet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1139,7 +355,7 @@ cell_t Native_FileSystemWatcher_RetryIntervalGet(SourcePawn::IPluginContext *con
 
 cell_t Native_FileSystemWatcher_RetryIntervalSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1159,7 +375,7 @@ cell_t Native_FileSystemWatcher_RetryIntervalSet(SourcePawn::IPluginContext *con
 
 cell_t Native_FileSystemWatcher_InternalBufferSizeGet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1171,7 +387,7 @@ cell_t Native_FileSystemWatcher_InternalBufferSizeGet(SourcePawn::IPluginContext
 
 cell_t Native_FileSystemWatcher_InternalBufferSizeSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1184,7 +400,7 @@ cell_t Native_FileSystemWatcher_InternalBufferSizeSet(SourcePawn::IPluginContext
 
 cell_t Native_FileSystemWatcher_OnStartedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1204,7 +420,7 @@ cell_t Native_FileSystemWatcher_OnStartedSet(SourcePawn::IPluginContext *context
 
 cell_t Native_FileSystemWatcher_OnStoppedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1224,7 +440,7 @@ cell_t Native_FileSystemWatcher_OnStoppedSet(SourcePawn::IPluginContext *context
 
 cell_t Native_FileSystemWatcher_OnCreatedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1244,7 +460,7 @@ cell_t Native_FileSystemWatcher_OnCreatedSet(SourcePawn::IPluginContext *context
 
 cell_t Native_FileSystemWatcher_OnDeletedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1264,7 +480,7 @@ cell_t Native_FileSystemWatcher_OnDeletedSet(SourcePawn::IPluginContext *context
 
 cell_t Native_FileSystemWatcher_OnModifiedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1284,7 +500,7 @@ cell_t Native_FileSystemWatcher_OnModifiedSet(SourcePawn::IPluginContext *contex
 
 cell_t Native_FileSystemWatcher_OnRenamedSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1304,7 +520,7 @@ cell_t Native_FileSystemWatcher_OnRenamedSet(SourcePawn::IPluginContext *context
 
 cell_t Native_FileSystemWatcher_IsWatchingGet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1316,7 +532,7 @@ cell_t Native_FileSystemWatcher_IsWatchingGet(SourcePawn::IPluginContext *contex
 
 cell_t Native_FileSystemWatcher_IsWatchingSet(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1339,7 +555,7 @@ cell_t Native_FileSystemWatcher_IsWatchingSet(SourcePawn::IPluginContext *contex
 
 cell_t Native_FileSystemWatcher_GetPath(SourcePawn::IPluginContext *context, const cell_t *params)
 {
-	FileSystemWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
+	SMDirectoryWatcher* watcher = g_FileSystemWatchers.GetWatcher(params[1]);
 	if (!watcher)
 	{
 		context->ReportError("Invalid FileSystemWatcher handle %x", params[1]);
@@ -1351,10 +567,10 @@ cell_t Native_FileSystemWatcher_GetPath(SourcePawn::IPluginContext *context, con
 
 	int bufferSize = params[3];
 
-	return watcher->GetRelativePath(buffer, bufferSize);
+	return watcher->GetPath(buffer, bufferSize);
 }
 
-sp_nativeinfo_s FileSystemWatcherManager::m_Natives[] =
+sp_nativeinfo_s SMDirectoryWatcherManager::m_Natives[] =
 {
 	{"FileSystemWatcher.FileSystemWatcher",	Native_FileSystemWatcher},
 	{"FileSystemWatcher.IsWatching.get", Native_FileSystemWatcher_IsWatchingGet},
